pyproject.toml
[project]
name = "todoapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.2",
    "django-extensions>=4.1",
    "django-stubs>=5.2.9",
    "django-stubs-ext>=5.2.9",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.29.0",
    "psycopg[binary]>=3.3.2",
    "python-dotenv>=1.2.1",
]
